Added example code to foreign_keys.applications package demonstrating how to work with linked entities (i.e. entities linked by foreign keys).
This code will read use the DAOs for Department and Employee (note: there is no Address DAO) to read the current content from the database into managed lists. If there are not enough departments currently in the system (the minimum required number is 5), the program will create up to 5 departments from a fixed set of names. It will then randomly generate a specified number of Addresses from a set of address components and use these to create up to 5 new Employees.
If there are at least 5 departments and at least 5 Employees persisted (i.e. in the database) already, it will just display the current state of the system.